fixing get request in backend. changing requestbody to request param in get mapping to adhere RESTful API principles.

+++ miners/tomorrow-miner/src/main/java/com/itay/weather/tomorrowminer/dto/ApiResponseValues.java
@@ -13,5 +13,5 @@ import lombok.Setter;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class ApiResponseValues {
     private double temperature;
-    private double humidity;
+    private int humidity;
 }
